renamed package structure to com.smartgridready. adapted to refactored libraries.
