+++ SampleCommunicator/src/main/java/com/smartgridready/communicator/example/AsynchronousSampleCommunicatorTest.java
@@ -1,22 +1,23 @@
-package ch.smartgridready.communicator.example;
+package com.smartgridready.communicator.example;
 
-import communicator.async.process.Parallel;
-import communicator.async.process.Processor;
-import communicator.async.process.ReadExec;
-import communicator.async.process.Sequence;
-import communicator.async.process.WriteExec;
-import communicator.common.api.values.Float32Value;
-import communicator.common.api.values.Float64Value;
-import communicator.common.api.values.StringValue;
-import communicator.common.api.values.Value;
-import communicator.common.runtime.GenDriverException;
-import communicator.common.runtime.GenDriverModbusException;
-import communicator.common.runtime.GenDriverSocketException;
-import communicator.modbus.impl.SGrModbusDevice;
-import communicator.rest.exception.RestApiAuthenticationException;
-import communicator.rest.exception.RestApiResponseParseException;
-import communicator.rest.exception.RestApiServiceCallException;
-import communicator.rest.impl.SGrRestApiDevice;
+import com.smartgridready.driver.api.common.GenDriverException;
+import com.smartgridready.driver.api.modbus.GenDriverModbusException;
+import com.smartgridready.driver.api.modbus.GenDriverSocketException;
+import com.smartgridready.communicator.async.process.Parallel;
+import com.smartgridready.communicator.async.process.Processor;
+import com.smartgridready.communicator.async.process.ReadExec;
+import com.smartgridready.communicator.async.process.Sequence;
+import com.smartgridready.communicator.async.process.WriteExec;
+import com.smartgridready.communicator.common.api.values.Float32Value;
+import com.smartgridready.communicator.common.api.values.Float64Value;
+import com.smartgridready.communicator.common.api.values.StringValue;
+import com.smartgridready.communicator.common.api.values.Value;
+import com.smartgridready.communicator.modbus.impl.SGrModbusDevice;
+import com.smartgridready.communicator.rest.exception.RestApiAuthenticationException;
+import com.smartgridready.communicator.rest.exception.RestApiResponseParseException;
+import com.smartgridready.communicator.rest.exception.RestApiServiceCallException;
+import com.smartgridready.communicator.rest.impl.SGrRestApiDevice;
+
 import org.awaitility.Awaitility;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
